scripts/heartbeat/requirements.txt
friday-logger==0.0.3
python-dotenv==1.0.1